refactor(apollo-vertex): split display tools into separate files
