+++ apps/apollo-vertex/locales/en.json
@@ -55,6 +55,7 @@
   "need_help": "Need help? Contact your administrator",
   "next": "Next",
   "next_slide": "Next slide",
+  "no_data_to_display": "No data to display.",
   "no_frameworks_found": "No frameworks found.",
   "no_results": "No results.",
   "of": "of",
