+++ apps/apollo-vertex/registry/ai-chat/charts/bar-chart-types.ts
@@ -1,0 +1,10 @@
+import type { ConfigFilter } from "./chart-shared-types";
+
+export interface BarChartConfig {
+  id: string;
+  name: string;
+  type: "bar";
+  dimensions: string[];
+  metrics: string[];
+  filters?: ConfigFilter[];
+}
