--- apps/apollo-vertex/registry/ai-chat/charts/chart-card.tsx
@@ -1,79 +1,0 @@
-"use client";
-
-import { useQueryClient } from "@tanstack/react-query";
-import {
-  BarChart,
-  type BarDataModel,
-  type DataAdapter,
-  TableChart,
-} from "@uipath/apollo-dashboarding";
-import { Card, CardContent } from "@/registry/card/card";
-import type { BarChartConfig, TableChartConfig } from "./chart-types";
-
-const DEFAULT_TABLE_STATE = { sortBy: null } as const;
-
-// ── Table ────────────────────────────────────────────────────────────────────
-
-interface TableChartCardProps {
-  configuration: TableChartConfig;
-  dataModel: {
-    id: string;
-    fields: Array<{
-      id: string;
-      display: string;
-      type: "numeric" | "string" | "boolean" | "datetime";
-    }>;
-  };
-  dataAdapter: DataAdapter;
-}
-
-export function TableChartCard({
-  configuration,
-  dataModel,
-  dataAdapter,
-}: TableChartCardProps) {
-  const queryClient = useQueryClient();
-
-  return (
-    <Card className="flex flex-col w-full h-[300px] gap-2 py-4">
-      <CardContent className="flex-1 overflow-hidden p-0">
-        <TableChart
-          configuration={configuration}
-          dataModel={dataModel}
-          dataAdapter={dataAdapter}
-          state={DEFAULT_TABLE_STATE}
-          queryClient={queryClient}
-        />
-      </CardContent>
-    </Card>
-  );
-}
-
-// ── Bar ──────────────────────────────────────────────────────────────────────
-
-interface BarChartCardProps {
-  configuration: BarChartConfig;
-  dataModel: BarDataModel;
-  dataAdapter: DataAdapter;
-}
-
-export function BarChartCard({
-  configuration,
-  dataModel,
-  dataAdapter,
-}: BarChartCardProps) {
-  const queryClient = useQueryClient();
-
-  return (
-    <Card className="flex flex-col w-full gap-2 py-4">
-      <CardContent className="flex-1 overflow-hidden p-0">
-        <BarChart
-          configuration={configuration}
-          dataModel={dataModel}
-          dataAdapter={dataAdapter}
-          queryClient={queryClient}
-        />
-      </CardContent>
-    </Card>
-  );
-}
