+++ apps/apollo-vertex/registry/ai-chat/charts/chart-shared-types.ts
@@ -1,0 +1,12 @@
+import type { BarChartConfiguration } from "@uipath/apollo-dashboarding";
+
+export type ColumnType = "numeric" | "string" | "boolean" | "datetime";
+
+export interface Column {
+  name: string;
+  type: ColumnType;
+}
+
+export type ConfigFilter = NonNullable<
+  BarChartConfiguration["filters"]
+>[number];
