+++ apps/apollo-vertex/registry/ai-chat/charts/no-data-message.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { useTranslation } from "react-i18next";
+
+export function NoDataMessage() {
+  const { t } = useTranslation();
+
+  return (
+    <p className="text-sm text-muted-foreground">{t("no_data_to_display")}</p>
+  );
+}
