+++ apps/apollo-vertex/registry/ai-chat/charts/table-chart-types.ts
@@ -1,0 +1,9 @@
+import type { ConfigFilter } from "./chart-shared-types";
+
+export interface TableChartConfig {
+  id: string;
+  name: string;
+  type: "table";
+  dimensions: string[];
+  filters?: ConfigFilter[];
+}
