+++ apps/apollo-vertex/registry/ai-chat/tools/barchart/index.ts
@@ -1,1 +1,5 @@
-export { barchartTool, BARCHART_TOOL_PROMPT } from "./barchart-tool";
+export {
+  createBarchartTool,
+  barchartTool,
+  BARCHART_TOOL_PROMPT,
+} from "./barchart-tool";
