+++ apps/apollo-vertex/registry/ai-chat/tools/choices/choices-tool.tsx
@@ -1,26 +1,20 @@
 "use client";
 
+import { toolDefinition } from "@tanstack/ai";
 import { z } from "zod";
-import { createDisplayTool } from "../tool-utils";
+import type { AiChatTool } from "../tool-types";
 import { AiChatSuggestions } from "../../components/ai-chat-suggestions";
 
 const choiceOptionSchema = z.object({
   id: z.string().describe("Unique identifier"),
   label: z.string().describe("Button text"),
   value: z.string().optional().describe("Optional payload"),
-  recommended: z
-    .boolean()
-    .optional()
-    .describe("Highlight as recommended"),
+  recommended: z.boolean().optional().describe("Highlight as recommended"),
 });
 
 const presentChoicesInput = z.object({
   prompt: z.string().describe("Short label above choices"),
-  options: z
-    .array(choiceOptionSchema)
-    .min(2)
-    .max(4)
-    .describe("2-4 options"),
+  options: z.array(choiceOptionSchema).min(2).max(4).describe("2-4 options"),
 });
 
 export type ChoiceOption = z.infer<typeof choiceOptionSchema>;
@@ -30,17 +24,23 @@ When the user asks for choices, options, or says things like "give me some choic
 Always mark exactly one option as recommended.
 After calling the tool keep your text reply short — the UI renders the options as buttons so do NOT repeat them in prose.`;
 
-export const choicesTool = createDisplayTool({
+const presentChoicesDef = toolDefinition({
   name: "presentChoices",
   description:
     "Present the user with 2-4 clickable choices. Call this tool whenever the user asks for choices, options, or wants to pick between alternatives. Mark one option as recommended when there is a clear best pick.",
   inputSchema: presentChoicesInput,
+});
+
+export const presentChoicesClient = presentChoicesDef.client((input) => input);
+
+export const choicesTool: AiChatTool = {
+  tool: presentChoicesClient,
   prompt: CHOICES_TOOL_PROMPT,
   render: (args, { onAction }) => (
     <AiChatSuggestions
-      prompt={args.prompt}
-      options={args.options}
-      onSelect={(option) => onAction(option.label)}
+      prompt={(args as z.infer<typeof presentChoicesInput>).prompt}
+      options={(args as z.infer<typeof presentChoicesInput>).options}
+      onSelect={(option: ChoiceOption) => onAction(option.label)}
     />
   ),
-});
+};
