+++ apps/apollo-vertex/registry/ai-chat/tools/choices/index.ts
@@ -1,1 +1,5 @@
-export { choicesTool, CHOICES_TOOL_PROMPT, type ChoiceOption } from "./choices-tool";
+export {
+  choicesTool,
+  CHOICES_TOOL_PROMPT,
+  type ChoiceOption,
+} from "./choices-tool";
