+++ apps/apollo-vertex/registry/ai-chat/tools/table/table-tool.tsx
@@ -1,14 +1,17 @@
 "use client";
 
+import { toolDefinition } from "@tanstack/ai";
+import type { DataAdapter } from "@uipath/apollo-dashboarding";
 import { Suspense } from "react";
 import { z } from "zod";
-import { createDisplayTool } from "../tool-utils";
-import { TableChartCard } from "../../charts/chart-card";
-import {
-  buildTableDataModel,
-  buildTableConfiguration,
-} from "../../charts/chart-data-model";
+import { NoDataMessage } from "../../charts/no-data-message";
 import { buildStaticDataAdapter } from "../../charts/static-data-adapter";
+import { TableChartCard } from "../../charts/table-chart-card";
+import {
+  buildTableConfiguration,
+  buildTableDataModel,
+} from "../../charts/table-data-model";
+import type { AiChatTool } from "../tool-types";
 
 const columnSchema = z.object({
   name: z.string().describe("Column name"),
@@ -26,38 +29,56 @@ const tableInput = z.object({
     .describe("Data rows as objects keyed by column name"),
 });
 
+type TableInput = z.infer<typeof tableInput>;
+
 export const TABLE_TOOL_PROMPT = `You have a "show_table" tool.
 When the user wants to see data as a table, call it with column definitions and row data.
 Each row is an object keyed by column name. Keep column count reasonable (3-8).
 After calling the tool, keep text reply short — the UI renders the table.`;
 
-export const tableTool = createDisplayTool({
+const showTableDef = toolDefinition({
   name: "show_table",
   description:
     "Display structured data as a table. Call this when the user wants to see data in tabular form. Provide column definitions with names and types, plus row data as objects keyed by column name.",
   inputSchema: tableInput,
-  prompt: TABLE_TOOL_PROMPT,
-  render: (args) => {
-    if (!args.columns?.length || !args.rows) {
-      return (
-        <p className="text-sm text-muted-foreground">
-          No data to display.
-        </p>
-      );
-    }
-    const rows = args.rows ?? [];
-    const dataModel = buildTableDataModel(args.columns);
-    const configuration = buildTableConfiguration(args.columns, args.title);
-    const cacheKey = JSON.stringify(rows).slice(0, 200);
-    const adapter = buildStaticDataAdapter(rows, cacheKey);
-    return (
-      <Suspense>
-        <TableChartCard
-          configuration={configuration}
-          dataModel={dataModel}
-          dataAdapter={adapter}
-        />
-      </Suspense>
-    );
-  },
 });
+
+export const showTableClient = showTableDef.client((input) => input);
+
+interface CreateTableToolOptions {
+  buildAdapter?: (
+    rows: Record<string, unknown>[],
+    cacheKey: string,
+  ) => DataAdapter;
+}
+
+export function createTableTool(options?: CreateTableToolOptions): AiChatTool {
+  const buildAdapter = options?.buildAdapter ?? buildStaticDataAdapter;
+
+  return {
+    tool: showTableClient,
+    prompt: TABLE_TOOL_PROMPT,
+    render: (args) => {
+      const { columns, rows, title } = args as TableInput;
+      if (!columns?.length || !rows) {
+        return <NoDataMessage />;
+      }
+      const safeRows = rows ?? [];
+      const dataModel = buildTableDataModel(columns);
+      const configuration = buildTableConfiguration(columns, title);
+      const cacheKey = JSON.stringify(safeRows).slice(0, 200);
+      const adapter = buildAdapter(safeRows, cacheKey);
+      return (
+        <Suspense>
+          <TableChartCard
+            configuration={configuration}
+            dataModel={dataModel}
+            dataAdapter={adapter}
+          />
+        </Suspense>
+      );
+    },
+  };
+}
+
+export const tableTool = createTableTool();
