+++ apps/apollo-vertex/registry/ai-chat/tools/tool-types.ts
@@ -1,9 +1,6 @@
 import type { AnyClientTool } from "@tanstack/ai";
 import type { ReactNode } from "react";
 
-/**
- * Associates a TanStack AI ClientTool with rendering and system prompt snippet.
- */
 export interface AiChatTool {
   tool: AnyClientTool;
   prompt: string;
@@ -13,7 +10,4 @@ export interface AiChatTool {
   ) => ReactNode;
 }
 
-/**
- * Record of display tools keyed by tool name, passed to AiChatMessage.
- */
 export type DisplayTools = Record<string, AiChatTool>;
