--- apps/apollo-vertex/registry/ai-chat/utils/ai-chat-utils.ts
@@ -1,2 +1,0 @@
-// Utility functions for ai-chat.
-// The findLatestChoices and ChoiceOption types have moved to tools/choices.
